refactor: Moved step helper and audit log to sub package.
