+++ src/main/java/com/synopsys/integration/detect/lifecycle/run/DetectRun.java
@@ -19,8 +19,8 @@ import com.synopsys.integration.detect.lifecycle.run.singleton.SingletonFactory;
 import com.synopsys.integration.detect.lifecycle.run.singleton.UtilitySingletons;
 import com.synopsys.integration.detect.lifecycle.run.step.IntelligentModeStepRunner;
 import com.synopsys.integration.detect.lifecycle.run.step.RapidModeStepRunner;
-import com.synopsys.integration.detect.lifecycle.run.step.StepHelper;
 import com.synopsys.integration.detect.lifecycle.run.step.UniversalStepRunner;
+import com.synopsys.integration.detect.lifecycle.run.step.utility.StepHelper;
 import com.synopsys.integration.detect.lifecycle.shutdown.ExitCodeManager;
 import com.synopsys.integration.detect.tool.UniversalToolsResult;
 import com.synopsys.integration.detect.tool.detector.factory.DetectorFactory;
