+++ src/main/java/com/synopsys/integration/detect/lifecycle/run/step/utility/StepHelper.java
@@ -5,7 +5,7 @@
  *
  * Use subject to the terms and conditions of the Synopsys End User Software License and Maintenance Agreement. All rights reserved worldwide.
  */
-package com.synopsys.integration.detect.lifecycle.run.step;
+package com.synopsys.integration.detect.lifecycle.run.step.utility;
 
 import java.io.IOException;
 import java.util.Optional;
